ARROW-10233: [Rust] Make array_value_to_string available in all Arrow builds
